moved PlayerCommand.java to API & fixed Tab Complete for IP, Find, Send and Server
