+++ api/src/main/java/net/md_5/bungee/api/plugin/PlayerCommand.java
@@ -1,13 +1,12 @@
-package net.md_5.bungee.command;
+package net.md_5.bungee.api.plugin;
 
 import com.google.common.base.Function;
 import com.google.common.base.Predicate;
 import com.google.common.collect.Iterables;
+
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
-import net.md_5.bungee.api.plugin.Command;
-import net.md_5.bungee.api.plugin.TabExecutor;
 
 public abstract class PlayerCommand extends Command implements TabExecutor
 {
