+++ proxy/src/main/java/net/md_5/bungee/command/CommandFind.java
@@ -1,10 +1,12 @@
 package net.md_5.bungee.command;
 
 import java.util.Collections;
+
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
+import net.md_5.bungee.api.plugin.PlayerCommand;
 
 public class CommandFind extends PlayerCommand
 {
