+++ proxy/src/main/java/net/md_5/bungee/command/CommandIP.java
@@ -4,9 +4,9 @@ import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
-import net.md_5.bungee.api.plugin.Command;
+import net.md_5.bungee.api.plugin.PlayerCommand;
 
-public class CommandIP extends Command
+public class CommandIP extends PlayerCommand
 {
 
     public CommandIP()
