+++ proxy/src/main/java/net/md_5/bungee/command/CommandSend.java
@@ -5,9 +5,9 @@ import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.config.ServerInfo;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
-import net.md_5.bungee.api.plugin.Command;
+import net.md_5.bungee.api.plugin.PlayerCommand;
 
-public class CommandSend extends Command
+public class CommandSend extends PlayerCommand
 {
 
     public CommandSend()
